Spell pointer offsets explicitly
Make the remaining cited pointer-offset calculations in elcore.c, inssubs.c, and pagevw.c explicit without changing the producing expressions or storage lifetimes.

The elcore debug print now subtracts from a char-buffer base, the FKP byte offset records its existing int narrowing explicitly, and the page-view DR deletion keeps the original PInPl recomputation while spelling the derived index as an int.

TODO.md now records that elxprocs.c:89 is already same-typed in this tree and that the live pointer-subtraction cleanup is complete for the cited sites.

Validation: git diff --check; rg -n 'pchBufStart - Global\(rgchBuf\)|bCur = hpch - \(char HUGE \*\)hpfkp|DeleteFromPl\(hpldr, pdr -' src/Opus (no matches); cmake --build out/macos-debug --target WORD1 -j2; ctest --test-dir out/macos-debug -R '^word1_port_smoke_test$|^opus_original_command_test$|^opus_word1_save_as_test$' --output-on-failure; cmake --build out/wasm-lp64-audit --target WORD1 -j2; ctest --test-dir out/wasm-lp64-audit -R '^opus_word1_save_as_test$|^opus_native_runtime_test$' --output-on-failure.

+++ src/Opus/wordtech/inssubs.c
@@ -649,7 +649,7 @@ int grpf;
 							!FNeHprgch((char HUGE *)pchProp, hpch + 1,
 							cchPropShare))
 						{ /* share existing property */
-						bCur = hpch - (char HUGE *)hpfkp;
+						bCur = (int)(hpch - (char HUGE *)hpfkp);
 						cchPropShare = 0;
 						break;  /* exit while */
 						}
